refactor: Ai버튼 항상 렌더링 되도록 수정
